requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.1
joblib==1.4.2
shap==0.46.0
matplotlib==3.10.0
seaborn==0.13.2
fpdf==1.7.2
google-generativeai==0.8.3
python-dotenv==1.0.0
Flask==2.0.1
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
python-socketio==5.12.1
python-engineio==4.11.2
bcrypt==4.3.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Werkzeug==2.0.1
SQLAlchemy==1.4.54